Add files via upload
The Quiz App is an Android application built using Java that allows users to test their knowledge on various topics. In this particular version, the quiz focuses on Indian politics in 2024. The app guides users through a series of multiple-choice questions, calculates their score, and provides feedback based on their performance